requirements.txt
torch
transformers
tqdm
Pillow